+++ console/src/app/pages/users/user-detail/auth-user-detail/auth-user-mfa/auth-user-mfa.component.ts
@@ -21,18 +21,6 @@ export function _base64ToArrayBuffer(base64: string): any {
     return bytes.buffer;
 }
 
-export function _arrayBufferToBase64(buffer: any): string {
-    let binary = '';
-    const bytes = new Uint8Array(buffer);
-    const len = bytes.byteLength;
-    for (let i = 0; i < len; i++) {
-        binary += String.fromCharCode(bytes[i]);
-    }
-    return btoa(binary).replace(/\+/g, '-')
-        .replace(/\//g, '_')
-        .replace(/=/g, '');
-}
-
 export interface WebAuthNOptions {
     challenge: string;
     rp: { name: string, id: string; };
